# src/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Delaware B2B API",
		"description": "Deze API is een voorbeeld van een B2B API voor een groothandel.",
		"version": "1.0",
		"contact": {
			"name": "Delaware",
			"url": "https://www.delaware.pro/"
		}
	},
	"servers": [{ "url": "http://localhost:8000/api" }],
	"tags": [
		{
			"name": "Aankopers"
		},
		{
			"name": "Bestellingen"
		},
		{
			"name": "Dozen"
		},
		{
			"name": "Klanten"
		},
		{
			"name": "Notificaties"
		},
		{
			"name": "Producten"
		},
		{
			"name": "Health"
		}
	],
	"paths": {
		"/aankopers/{id}": {
			"get": {
				"tags": ["Aankopers"],
				"description": "Haal een aankoper op",
				"operationId": "getAankoperById",
				"parameters": [
					{
						"$ref": "#/components/parameters/id"
					}
				],
				"responses": {
					"200": {
						"description": "Aankoper gevonden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Aankoper"
								}
							}
						}
					},
					"400": {
						"description": "Ongeldige aankoperId"
					},
					"404": {
						"description": "Aankoper niet gevonden"
					}
				}
			}
		},
		"/bestellingen": {
			"post": {
				"tags": ["Bestellingen"],
				"description": "Maak een nieuwe bestelling aan",
				"operationId": "postBestelling",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Bestelling"
							}
						}
					}
				}
			}
		},
		"/bestellingen/{id}": {
			"get": {
				"tags": ["Bestellingen"],
				"description": "Haal de details van een bestelling op",
				"operationId": "getBestellingById",
				"parameters": [
					{
						"$ref": "#/components/parameters/id"
					}
				],
				"responses": {
					"200": {
						"description": "Bestelling gevonden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Bestelling"
								}
							}
						}
					},
					"400": {
						"description": "Ongeldige bestellingId"
					},
					"404": {
						"description": "Bestelling niet gevonden"
					}
				}
			},
			"put": {
				"tags": ["Bestellingen"],
				"description": "Update een bestelling",
				"operationId": "putBestelling",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Bestelling"
							}
						}
					}
				}
			}
		},
		"/bestellingen/{id}/{trackAndTraceCode}": {
			"get": {
				"tags": ["Bestellingen"],
				"description": "Haal de details van een bestelling op adhv. een trackAndTraceCode (Non-Protected Endpoint)",
				"operationId": "getBestellingByTTCode",
				"parameters": [
					{
						"$ref": "#/components/parameters/id"
					},
					{
						"name": "trackAndTraceCode",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Bestelling gevonden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Bestelling"
								}
							}
						}
					},
					"400": {
						"description": "Ongeldige bestellingId"
					},
					"404": {
						"description": "Bestelling niet gevonden"
					},
					"401": {
						"description": "Foute combinatie trackAndTraceCode en bestelling"
					}
				}
			}
		},
		"/bestellingen/klanten/{id}": {
			"get": {
				"tags": ["Klanten", "Bestellingen"],
				"description": "Haal alle bestellingen van een klant op",
				"operationId": "getBestellingenByKlantId",
				"parameters": [
					{
						"$ref": "#/components/parameters/id"
					}
				],
				"responses": {
					"200": {
						"description": "Bestellingen gevonden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BestellingLijst"
								}
							}
						}
					},
					"400": {
						"description": "Ongeldige klantId"
					},
					"404": {
						"description": "Klant niet gevonden"
					}
				}
			}
		},
		"/dozen": {
			"get": {
				"tags": ["Dozen"],
				"description": "Haal alle dozen op",
				"operationId": "getDozen",
				"responses": {
					"200": {
						"description": "Dozen gevonden",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Doos"
									}
								}
							}
						}
					}
				}
			}
		},
		"/klanten/{id}": {
			"get": {
				"tags": ["Klanten"],
				"description": "Haal een klant en al zijn aankopers op",
				"operationId": "getKlantById",
				"parameters": [
					{
						"$ref": "#/components/parameters/id"
					}
				],
				"responses": {
					"200": {
						"description": "Klant gevonden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Klant"
								}
							}
						}
					},
					"400": {
						"description": "Ongeldige klantId"
					},
					"404": {
						"description": "Klant niet gevonden"
					}
				}
			}
		},
		"/notificaties/{id}": {
			"put": {
				"tags": ["Notificaties"],
				"description": "Update een notificatie",
				"operationId": "putNotificatie",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Notificatie"
							}
						}
					}
				}
			}
		},
		"/notificaties/klanten/{id}": {
			"get": {
				"tags": ["Klanten", "Notificaties"],
				"description": "Haal alle notificaties van een klant op",
				"operationId": "getNotificatiesByKlantId",
				"parameters": [
					{
						"$ref": "#/components/parameters/id"
					}
				],
				"responses": {
					"200": {
						"description": "Notificaties gevonden",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Notificatie"
									}
								}
							}
						}
					},
					"400": {
						"description": "Ongeldige klantId"
					},
					"404": {
						"description": "Klant niet gevonden"
					}
				}
			}
		},
		"/producten": {
			"get": {
				"tags": ["Producten"],
				"description": "Haal een pagina producten op (8 producten per pagina)",
				"operationId": "getProducten",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"description": "Paginanummer",
						"required": false,
						"schema": {
							"type": "integer",
							"default": 1
						}
					}
				],
				"responses": {
					"200": {
						"description": "Producten gevonden",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Product"
									}
								}
							}
						}
					}
				}
			}
		},
		"/producten/{id}": {
			"get": {
				"tags": ["Producten"],
				"description": "Haal een product op",
				"operationId": "getProductById",
				"parameters": [
					{
						"$ref": "#/components/parameters/id"
					}
				],
				"responses": {
					"200": {
						"description": "Product gevonden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					},
					"400": {
						"description": "Ongeldige productId"
					},
					"404": {
						"description": "Product niet gevonden"
					}
				}
			}
		},
		"/health": {
			"get": {
				"tags": ["Health"],
				"description": "Haal de status van de applicatie op",
				"operationId": "getHealth",
				"responses": {
					"200": {
						"description": "Applicatie is gezond"
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"Aankoper": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"naam": {
						"type": "string"
					},
					"voornaam": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"klant": {
						"$ref": "#/components/schemas/Klant"
					}
				}
			},
			"Adres": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"straat": {
						"type": "string"
					},
					"huisnummer": {
						"type": "string"
					},
					"postcode": {
						"type": "string"
					},
					"land": {
						"type": "string"
					}
				}
			},
			"Bestelling": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"status": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"trackAndTrace": {
						"type": "object",
						"properties": {
							"code": {
								"type": "string"
							}
						}
					},
					"leverAdres": {
						"$ref": "#/components/schemas/Adres"
					},
					"doos": {
						"type": "object",
						"properties": {
							"naam": {
								"type": "string"
							}
						}
					},
					"bestelregels": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Bestelregel"
						}
					},
					"totaalPrijs": {
						"type": "number"
					}
				}
			},
			"BestellingLijst": {
				"type": "array",
				"items": {
					"type": "object",
					"properties": {
						"id": {
							"type": "integer"
						},
						"status": {
							"type": "integer"
						},
						"createdAt": {
							"type": "string",
							"format": "date-time"
						},
						"totaalPrijs": {
							"type": "number"
						}
					}
				}
			},
			"Bestelregel": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"aantal": {
						"type": "integer"
					},
					"product": {
						"$ref": "#/components/schemas/Product"
					}
				}
			},
			"Doos": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"naam": {
						"type": "string"
					},
					"actief": {
						"type": "boolean"
					},
					"breedte": {
						"type": "number"
					},
					"hoogte": {
						"type": "number"
					},
					"lengte": {
						"type": "number"
					},
					"prijs": {
						"type": "number"
					},
					"type": {
						"type": "number"
					}
				}
			},
			"Klant": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"naam": {
						"type": "string"
					},
					"logo": {
						"type": "string"
					},
					"adres": {
						"$ref": "#/components/schemas/Adres"
					},
					"gsm": {
						"type": "string"
					},
					"aankopers": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Aankoper"
						}
					}
				}
			},
			"Notificatie": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"bestelling": {
						"type": "object",
						"properties": {
							"id": {
								"type": "integer"
							},
							"status": {
								"type": "string"
							}
						}
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"foto": {
						"type": "string"
					},
					"gelezen": {
						"type": "boolean"
					}
				}
			},
			"Product": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"naam": {
						"type": "string"
					},
					"foto": {
						"type": "string"
					},
					"omschrijving": {
						"type": "string"
					},
					"prijs": {
						"type": "integer"
					},
					"stock": {
						"type": "integer"
					},
					"levertijd": {
						"type": "integer"
					}
				}
			}
		},
		"parameters": {
			"id": {
				"name": "id",
				"in": "path",
				"required": true,
				"schema": {
					"type": "integer"
				}
			}
		}
	}
}